fix integer overflow

+++ gdb/factorial.c
@@ -4,24 +4,24 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-int factorial(int n);
+long long factorial(int n);
 
 int main(void) {
 	
 	int n = 5;
-	int f = factorial(n);
+	long long f = factorial(n);
 	printf("The factorial of %d is %d.\n", n, f);
 	n = 17;
 	f = factorial(n);
-	printf("The factorial of %d is %d.\n", n, f);
+	printf("The factorial of %d is %lu.\n", n, f);
 
 	return 0;
 		
 }
 //A factorial is calculated by n! = n * (n - 1) * (n - 2) * ... * 1
 //E.g. 5! = 5 * 4 * 3 * 2 * 1 = 120
-int factorial(int n) {
-	int f = 1;
+long long factorial(int n) {
+	long long f = 1;
 	int i = 1;
 	while (i <= n) {
 		f = f * i;
